+++ src/engine/client/application/Direct3d9/src/win32/Direct3d9_StaticVertexBufferData.cpp
@@ -128,7 +128,7 @@ void Direct3d9_StaticVertexBufferData::unlock()
 
 int Direct3d9_StaticVertexBufferData::getSortKey()
 {
-	return reinterpret_cast<int>(m_d3dVertexBuffer);
+	return static_cast<int>(reinterpret_cast<intptr_t>(m_d3dVertexBuffer));
 }
 
 // ----------------------------------------------------------------------
